requirements.txt
mysql-connector-python>=8.0.0
Flask>=2.0.0

